requirements.txt
numpy==1.18.5
scipy==1.5.1
pytz==2020.1
psycopg2==2.8.5
Pyrebase==3.0.27
imutils==0.5.3
Django==2.2.12
django_crispy_forms==1.9.2
Pillow==7.2.0
dj-database-url==0.5.0
opencv-python-headless
gunicorn
django-heroku